Add main window
