+++ src/window/mod.rs
@@ -1,0 +1,2 @@
+mod main;
+pub use main::*;
